fix: :recycle: moved bullet apply_velocity to enemies
Issue: #6

敵の弾の移動処理を統一
